backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
pydantic==2.7.0
python-dotenv==1.0.1
httpx==0.27.0
chromadb==0.5.3
PyYAML==6.0.1